Changed calendar task to a model
still need to sort out errors, edge cases and maybe make it a singleton
