--- app/src/main/java/com/hva/m2mobi/m2hva_reservationsystem/utils/CalendarEventListener.java
@@ -1,7 +1,0 @@
-package com.hva.m2mobi.m2hva_reservationsystem.utils;
-
-import com.google.api.services.calendar.model.Events;
-
-public interface CalendarEventListener {
-    void onCalendarEventsReturned(Events events);
-}
